Create useQuiz Hook
