+++ src/pages/Questionnaire.tsx
@@ -1,16 +1,11 @@
-import React, { useState, useEffect, useCallback } from 'react';
+import React from 'react';
 import Question from '../components/Question';
 import Error500 from '../pages/Error500';
-import { useQuestions } from '../hooks/useQuestions';
-import { TQuestion } from '../types/types';
 import Loader from '../components/Loader';
 import PageWrapper from '../components/PageWrapper';
 import { makeStyles, Grid, LinearProgress, Box } from '@material-ui/core';
-import GetZipCode from './GetZipCode';
-import { TAnswers } from '../types/types';
-import { useResponses } from '../hooks/useResponses';
 import PrevButton from '../components/PrevButton';
-import { pushQuestionToDataLayer } from '../analytics';
+import { useQuiz } from '../hooks/useQuiz';
 
 const styles = makeStyles({
   root: {
@@ -38,113 +33,15 @@ const styles = makeStyles({
 
 const Questionaire: React.FC<{}> = () => {
   const classes = styles();
-  const { questions, questionsLoading, questionsError } = useQuestions();
-  // List of answers
-
-  const [answers, setAnswers] = useState<TAnswers | null>(null);
-  // Questions still to be answered
-
-  const [questionsToAnswer, setQuestionsToAnswer] = useState<TQuestion[] | []>(
-    []
-  );
-  // Questions which the user has already answered (allowing us to do back)
-  const [questionsAnswered, setQuestionsAnswered] = useState<TQuestion[] | []>(
-    []
-  );
-  // Current Question being answered
-  const [currentQuestion, setCurrentQuestion] = useState<TQuestion | null>(
-    null
-  );
-  // Question number user is on
-  const [progress, setProgress] = useState(0);
-
-  // Update the state of responses
-  const { dispatch } = useResponses();
-
-  // Move forward a question
-  const changeQuestionForward = useCallback(() => {
-    // The questionnaire always presents the user with the last question on the questionsToAnswer array. When the question is answered it is popped from the array and then pushed on to the questionsAnswered array. This is to allow us to go back in future.
-    const oldCurrentQuestion = questionsToAnswer?.pop();
-    const updatedQuestionsAnswered = [...questionsAnswered];
-    if (oldCurrentQuestion) {
-      updatedQuestionsAnswered.push(oldCurrentQuestion);
-      setQuestionsAnswered(updatedQuestionsAnswered);
-      setQuestionsToAnswer(questionsToAnswer);
-    }
-    // Set a new currentQuestion to the last question in the list
-    setCurrentQuestion(questionsToAnswer[questionsToAnswer.length]);
-    setProgress(progress + 1);
-  }, [setQuestionsToAnswer, questionsToAnswer, questionsAnswered, progress]);
-
-  const changeQuestionBackward = useCallback(() => {
-    if (questionsAnswered.length < 1) {
-      return;
-    }
-    const updatedAnswered = [...questionsAnswered];
-    const updatedToAnswer = [...questionsToAnswer];
-    const curr = updatedAnswered.pop();
-    if (curr) {
-      updatedToAnswer?.push(curr);
-      setQuestionsToAnswer(updatedToAnswer);
-      setQuestionsAnswered(updatedAnswered);
-    }
-    // Set a new currentQuestion to the last question in the list...
-    setCurrentQuestion(questionsToAnswer[questionsToAnswer.length]);
-    setProgress(progress - 1);
-  }, [setQuestionsToAnswer, questionsToAnswer, questionsAnswered, progress]);
-
-  // Handle answering of a question
-  const setAnswer = (questionId: number, answerId: string) => {
-    // Saving answer to state
-    dispatch({
-      type: 'ADD_SETONE',
-      action: { questionId: questionId, answerId: parseInt(answerId) },
-    });
-    changeQuestionForward();
-  };
-
-  // Setting the questions on load;
-  useEffect(() => {
-    // TODO - For just now we are only using SetOne
-    if (questions.SetOne) {
-      const questionsToAnswer: TQuestion[] = [...questions.SetOne];
-      setQuestionsToAnswer(questionsToAnswer);
-    }
-  }, [questions]);
-
-  // Setting the answers on load
-  useEffect(() => {
-    if (questions.Answers && !answers) {
-      setAnswers(questions.Answers);
-    }
-  }, [questions, answers, questionsToAnswer, setQuestionsToAnswer]);
-
-  // Set the first question to answer
-  useEffect(() => {
-    if (!currentQuestion && questionsToAnswer.length) {
-      // Set to the last question of the array
-
-      const currentQuestion = questionsToAnswer[questionsToAnswer.length - 1];
-      setCurrentQuestion(currentQuestion);
-    }
-  }, [
-    questionsToAnswer,
+  const {
     currentQuestion,
-    changeQuestionForward,
-    setCurrentQuestion,
-  ]);
-
-  // add question id to url (for tracking)
-  useEffect(() => {
-    if (currentQuestion) {
-      pushQuestionToDataLayer(currentQuestion.id);
-    }
-  }, [currentQuestion]);
-
-  //Show submit page when quiz is complete - This just a hack just now to show the quiz is completed, we need a better machanism in future.
-  if (progress === 10) {
-    return <GetZipCode />;
-  }
+    answers,
+    progress,
+    questionsError,
+    questionsLoading,
+    setAnswer,
+    changeQuestionBackward,
+  } = useQuiz();
 
   if (questionsError) {
     return <Error500 />;
